Move notification code out of service
