+++ async/src/main/java/org/odk/collect/async/CoroutineAndWorkManagerScheduler.kt
@@ -8,35 +8,13 @@ import androidx.work.OneTimeWorkRequest
 import androidx.work.PeriodicWorkRequest
 import androidx.work.WorkInfo
 import androidx.work.WorkManager
-import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.Dispatchers
-import kotlinx.coroutines.Runnable
-import kotlinx.coroutines.cancel
-import kotlinx.coroutines.delay
-import kotlinx.coroutines.isActive
-import kotlinx.coroutines.launch
-import kotlinx.coroutines.withContext
 import java.util.concurrent.TimeUnit
-import java.util.function.Consumer
-import java.util.function.Supplier
 import kotlin.coroutines.CoroutineContext
 
-class CoroutineAndWorkManagerScheduler(private val foregroundContext: CoroutineContext, private val backgroundContext: CoroutineContext, private val workManager: WorkManager) : Scheduler {
+class CoroutineAndWorkManagerScheduler(foregroundContext: CoroutineContext, backgroundContext: CoroutineContext, private val workManager: WorkManager) : CoroutineScheduler(foregroundContext, backgroundContext) {
 
     constructor(workManager: WorkManager) : this(Dispatchers.Main, Dispatchers.IO, workManager) // Needed for Java construction
-
-    override fun <T> immediate(background: Supplier<T>, foreground: Consumer<T>) {
-        CoroutineScope(foregroundContext).launch {
-            val result = withContext(backgroundContext) { background.get() }
-            foreground.accept(result)
-        }
-    }
-
-    override fun immediate(foreground: java.lang.Runnable) {
-        CoroutineScope(foregroundContext).launch {
-            foreground.run()
-        }
-    }
 
     override fun networkDeferred(tag: String, spec: TaskSpec) {
         val constraints = Constraints.Builder()
@@ -65,19 +43,6 @@ class CoroutineAndWorkManagerScheduler(private val foregroundContext: CoroutineC
         workManager.enqueueUniquePeriodicWork(tag, ExistingPeriodicWorkPolicy.REPLACE, workRequest)
     }
 
-    override fun repeat(foreground: Runnable, repeatPeriod: Long): Cancellable {
-        val repeatScope = CoroutineScope(foregroundContext)
-
-        repeatScope.launch {
-            while (isActive) {
-                foreground.run()
-                delay(repeatPeriod)
-            }
-        }
-
-        return ScopeCancellable(repeatScope)
-    }
-
     override fun cancelDeferred(tag: String) {
         workManager.cancelUniqueWork(tag)
     }
@@ -95,13 +60,5 @@ class CoroutineAndWorkManagerScheduler(private val foregroundContext: CoroutineC
         }
 
         return false
-    }
-}
-
-private class ScopeCancellable(private val scope: CoroutineScope) : Cancellable {
-
-    override fun cancel(): Boolean {
-        scope.cancel()
-        return true
     }
 }
