+++ audiorecorder/src/main/java/org/odk/collect/audiorecorder/DaggerSetup.kt
@@ -3,12 +3,12 @@ package org.odk.collect.audiorecorder
 import android.app.Application
 import android.content.Context
 import android.media.MediaRecorder
-import androidx.work.WorkManager
 import dagger.BindsInstance
 import dagger.Component
 import dagger.Module
 import dagger.Provides
-import org.odk.collect.async.CoroutineAndWorkManagerScheduler
+import kotlinx.coroutines.Dispatchers
+import org.odk.collect.async.CoroutineScheduler
 import org.odk.collect.async.Scheduler
 import org.odk.collect.audiorecorder.recorder.MediaRecorderRecorder
 import org.odk.collect.audiorecorder.recorder.RealMediaRecorderWrapper
@@ -76,7 +76,7 @@ internal open class AudioRecorderDependencyModule {
 
     @Provides
     open fun providesScheduler(application: Application): Scheduler {
-        return CoroutineAndWorkManagerScheduler(WorkManager.getInstance(application))
+        return CoroutineScheduler(Dispatchers.Main, Dispatchers.IO)
     }
 }
